Extract frontend executable functionality to a separate contextual class. (#7849)
Fixes #7771
# Conflicts:
#	flow-server/src/main/java/com/vaadin/flow/server/DevModeHandler.java
#	flow-server/src/main/java/com/vaadin/flow/server/frontend/FrontendUtils.java
#	flow-server/src/test/java/com/vaadin/flow/server/frontend/TaskRunPnpmInstallTest.java
