+++ flow-server/src/main/java/com/vaadin/flow/server/DevModeHandler.java
@@ -42,6 +42,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.vaadin.flow.function.DeploymentConfiguration;
+import com.vaadin.flow.server.frontend.FrontendTools;
 import com.vaadin.flow.server.frontend.FrontendUtils;
 
 import static com.vaadin.flow.server.Constants.SERVLET_PARAMETER_DEVMODE_TRANSPILE;
@@ -56,7 +57,6 @@ import static com.vaadin.flow.server.frontend.FrontendUtils.WEBPACK_CONFIG;
 import static com.vaadin.flow.server.frontend.FrontendUtils.YELLOW;
 import static com.vaadin.flow.server.frontend.FrontendUtils.commandToString;
 import static com.vaadin.flow.server.frontend.FrontendUtils.console;
-import static com.vaadin.flow.server.frontend.FrontendUtils.getNodeExecutable;
 import static java.net.HttpURLConnection.HTTP_NOT_FOUND;
 import static java.net.HttpURLConnection.HTTP_OK;
 
@@ -142,17 +142,18 @@ public final class DevModeHandler {
         ProcessBuilder processBuilder = new ProcessBuilder()
                 .directory(npmFolder);
 
-        FrontendUtils.validateNodeAndNpmVersion(npmFolder.getAbsolutePath());
+        FrontendTools tools = new FrontendTools(npmFolder.getAbsolutePath(),
+                () -> FrontendUtils.getVaadinHomeDirectory().getAbsolutePath());
+        tools.validateNodeAndNpmVersion();
 
         boolean useHomeNodeExec = config.getBooleanProperty(
                 Constants.REQUIRE_HOME_NODE_EXECUTABLE, false);
 
         String nodeExec = null;
         if (useHomeNodeExec) {
-            nodeExec = FrontendUtils
-                    .ensureNodeExecutableInHome(npmFolder.getAbsolutePath());
+            nodeExec = tools.forceAlternativeNodeExecutable();
         } else {
-            nodeExec = getNodeExecutable(npmFolder.getAbsolutePath());
+            nodeExec = tools.getNodeExecutable();
         }
 
         List<String> command = new ArrayList<>();
